update gccgo support

+++ src/hotspin-core/gpu/brillouin-wrapper.go
@@ -25,11 +25,11 @@ func BrillouinAsync(msat0 *Array, msat0T0 *Array, T *Array, Tc *Array, S *Array,
 
 	// Calling the CUDA functions
 	C.brillouinAsync(
-		(*C.float)(unsafe.Pointer((msat0.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((msat0T0.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((T.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((Tc.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((S.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(msat0.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(msat0T0.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(T.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(Tc.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(S.Comp[X].pointer))),
 
 		(C.float)(msat0Mul),
 		(C.float)(msat0T0Mul),
@@ -37,5 +37,5 @@ func BrillouinAsync(msat0 *Array, msat0T0 *Array, T *Array, Tc *Array, S *Array,
 		(C.float)(SMul),
 
 		(C.int)(msat0.partLen3D),
-		(C.CUstream)(unsafe.Pointer((stream))))
+		(C.CUstream)(unsafe.Pointer(uintptr(stream))))
 }
