+++ src/hotspin-core/gpu/wrap_cp.go
@@ -25,13 +25,13 @@ func CpAsync(cp *Array, T *Array, Td *Array, n *Array, TdMul float64, stream Str
 
 	// Calling the CUDA functions
 	C.cpAsync(
-		(*C.float)(unsafe.Pointer((cp.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((T.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((Td.Comp[X].pointer))),
-		(*C.float)(unsafe.Pointer((n.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(cp.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(T.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(Td.Comp[X].pointer))),
+		(*C.float)(unsafe.Pointer(uintptr(n.Comp[X].pointer))),
 
 		(C.float)(TdMul),
 
 		(C.int)(cp.partLen3D),
-		(C.CUstream)(unsafe.Pointer((stream))))
+		(C.CUstream)(unsafe.Pointer(uintptr(stream))))
 }
